producer/requirements.txt
kafka-python==2.0.1
faker==4.1.1